Extract statement generation to separate class
