+++ src/main/scala/com/abc/Customer.scala
@@ -20,32 +20,5 @@ class Customer(val name: String, var accounts: ListBuffer[Account] = ListBuffer(
     from.withdraw(amount)
     to.deposit(amount)
   }
-
-  /**
-   * This method gets a statement
-   */
-  def getStatement: String = {
-    val totalAcrossAllAccounts = accounts.map(_.balance).sum
-    val statement = f"Statement for $name\n" +
-      accounts.map(statementForAccount).mkString("\n", "\n\n", "\n") +
-      s"\nTotal In All Accounts ${toDollars(totalAcrossAllAccounts)}"
-    statement
-  }
-
-  private def statementForAccount(a: Account): String = {
-    val transactionSummary = a.transactions.map(t => withdrawalOrDepositText(t) + " " + toDollars(t.amount.abs))
-      .mkString("  ", "\n  ", "\n")
-    val totalSummary = s"Total ${toDollars(a.transactions.map(_.amount).sum)}"
-    a.name + " Account\n" + transactionSummary + totalSummary
-  }
-
-  private def withdrawalOrDepositText(t: Transaction) =
-    t.amount match {
-      case a if a < 0 => "withdrawal"
-      case a if a > 0 => "deposit"
-      case _ => "N/A"
-    }
-
-  private def toDollars(number: Double): String = f"$$$number%.2f"
 }
